# be/docs/swagger.yaml
basePath: /api
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.ErrorData'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  models.ErrorData:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
  models.ExportResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      download_url:
        type: string
      file_size:
        type: integer
      job_id:
        type: string
      record_count:
        type: integer
      status:
        description: '"pending", "processing", "completed", "failed"'
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@form.com
    name: API Support
  description: API for form application with JWT authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: form API
  version: "1.0"
paths:
...
  /exports/{job_id}/download:
    get:
      consumes:
      - application/json
      description: Download the exported file when the job is completed
      parameters:
      - description: Export job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Download exported file
      tags:
      - export
  /exports/{job_id}/status:
    get:
      consumes:
      - application/json
      description: Check the status of an export job and get download URL when ready
      parameters:
      - description: Export job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ExportResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get export job status
      tags:
      - export
...
  /forms/{form_id}/export:
    get:
      consumes:
      - application/json
      description: Export form responses as CSV, Excel, or JSON with customizable
        columns and data inclusion
      parameters:
      - description: Form ID
        in: path
        name: form_id
        required: true
        type: string
      - description: Export format
        enum:
        - csv
        - excel
        - json
        in: query
        name: format
        required: true
        type: string
      - collectionFormat: csv
        description: Specific columns to include
        in: query
        items:
          type: string
        name: columns
        type: array
      - description: Start date filter (YYYY-MM-DD)
        in: query
        name: from_date
        type: string
      - description: End date filter (YYYY-MM-DD)
        in: query
        name: to_date
        type: string
      - default: true
        description: Include detailed answers
        in: query
        name: include_answers
        type: boolean
      - default: false
        description: Include analytics summary
        in: query
        name: include_analytics
        type: boolean
      - description: Filter by pass/fail status
        in: query
        name: is_passed
        type: boolean
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Export form responses in various formats
      tags:
      - export
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"